refactor: remove settings and subscription pages
