--- pages/dashboard/settings.tsx
@@ -1,6 +1,0 @@
-import { dashboardSettingsPage } from '@application';
-
-const { getServerSideProps, SettingsPage } = dashboardSettingsPage;
-
-// eslint-disable-next-line no-restricted-exports
-export { SettingsPage as default, getServerSideProps };
