--- pages/dashboard/subscription.tsx
@@ -1,6 +1,0 @@
-import { dashboardSubscriptionPage } from '@application';
-
-const { getServerSideProps, SubscriptionPage } = dashboardSubscriptionPage;
-
-// eslint-disable-next-line no-restricted-exports
-export { SubscriptionPage as default, getServerSideProps };
