+++ src/application/pages/dashboard/index.ts
@@ -2,5 +2,3 @@ export * as dashboardLikesPage from './likes/likes.page';
 export * as dashboardProfilePage from './profile/profile.page';
 export * as dashboardMatchesPage from './matches/matches.page';
 export * as dashboardDiscoverPage from './discover/discover.page';
-export * as dashboardSettingsPage from './settings/settings.page';
-export * as dashboardSubscriptionPage from './subscription/subscription.page';
