+++ src/application/pages/dashboard/layouts/desktop-dashboard/desktop-dashboard.defaults.tsx
@@ -1,7 +1,5 @@
 import {
-  CardMembership as CardMembershipIcon,
   JoinInner as JoinInnerIcon,
-  Settings as SettingsIcon,
   StarBorder as StarBorderIcon,
   Style as StyleIcon,
 } from '@mui/icons-material';
@@ -13,8 +11,6 @@ export const translationStrings = [
   'component.dashboard.sidemenu.option.discover',
   'component.dashboard.sidemenu.option.matches',
   'component.dashboard.sidemenu.option.likes',
-  'component.dashboard.sidemenu.option.subscription',
-  'component.dashboard.sidemenu.option.settings',
 ] as const;
 
 export const getSideMenuOptions = (translations: Record<string, string>) => [
@@ -32,15 +28,5 @@ export const getSideMenuOptions = (translations: Record<string, string>) => [
     icon: <StarBorderIcon />,
     route: ERoutes.DASHBOARD_LIKES,
     label: translations.componentDashboardSidemenuOptionLikes,
-  },
-  {
-    icon: <CardMembershipIcon />,
-    route: ERoutes.DASHBOARD_SUBSCRIPTION,
-    label: translations.componentDashboardSidemenuOptionSubscription,
-  },
-  {
-    icon: <SettingsIcon />,
-    route: ERoutes.DASHBOARD_SETTINGS,
-    label: translations.componentDashboardSidemenuOptionSettings,
   },
 ];
