+++ src/application/pages/dashboard/layouts/mobile-dashboard/mobile-dashboard.defaults.tsx
@@ -1,6 +1,5 @@
 import {
   JoinInner as JoinInnerIcon,
-  Settings as SettingsIcon,
   StarBorder as StarBorderIcon,
   Style as StyleIcon,
 } from '@mui/icons-material';
@@ -11,7 +10,6 @@ export const translationStrings = [
   'component.dashboard.sidemenu.option.discover',
   'component.dashboard.sidemenu.option.matches',
   'component.dashboard.sidemenu.option.likes',
-  'component.dashboard.sidemenu.option.settings',
 ] as const;
 
 export const getBottomNavigationOptions = (translations: Record<string, string>) => [
@@ -29,10 +27,5 @@ export const getBottomNavigationOptions = (translations: Record<string, string>)
     icon: <StarBorderIcon />,
     route: ERoutes.DASHBOARD_LIKES,
     label: translations.componentDashboardSidemenuOptionLikes,
-  },
-  {
-    icon: <SettingsIcon />,
-    route: ERoutes.DASHBOARD_SETTINGS,
-    label: translations.componentDashboardSidemenuOptionSettings,
   },
 ];
