+++ src/application/pages/dashboard/profile/profile.types.ts
@@ -1,1 +1,5 @@
-export interface IProfileProps {}
+import { ISubscriptionPlan } from '@interfaces';
+
+export interface IProfileProps {
+  plans: ISubscriptionPlan[];
+}
