--- src/application/pages/dashboard/settings/settings.page.tsx
@@ -1,34 +1,0 @@
-import { GetServerSideProps } from 'next';
-
-import { Settings } from '@domain';
-
-import { hasOwnProperty, useDeviceDetect } from '@utils';
-
-import { DesktopDashboardLayout, MobileDashboardLayout } from '../layouts';
-import { verifyUserAccess } from '../utils';
-
-export const SettingsPage = (): JSX.Element => {
-  const { deviceData } = useDeviceDetect();
-
-  const DashboardLayout = deviceData.isSmallerThanLG
-    ? MobileDashboardLayout
-    : DesktopDashboardLayout;
-
-  return (
-    <DashboardLayout>
-      <Settings />
-    </DashboardLayout>
-  );
-};
-
-export const getServerSideProps = async (serverSideProps: GetServerSideProps) => {
-  const result = await verifyUserAccess(serverSideProps);
-
-  if (!hasOwnProperty(result, 'profileData') || !hasOwnProperty(result, 'user')) {
-    return result;
-  }
-
-  return {
-    props: {},
-  };
-};
