--- src/application/pages/dashboard/settings/settings.types.ts
@@ -1,1 +1,0 @@
-export interface ISettingsProps {}
