--- src/application/pages/dashboard/subscription/subscription.page.tsx
@@ -1,62 +1,0 @@
-import { GetServerSideProps } from 'next';
-import { Stripe } from 'stripe';
-
-import { Subscription } from '@domain';
-
-import { hasOwnProperty, useDeviceDetect } from '@utils';
-
-import { DesktopDashboardLayout, MobileDashboardLayout } from '../layouts';
-import { verifyUserAccess } from '../utils';
-import { ISubscriptionProps } from './subscription.types';
-
-const stripeSecretKey = process.env.STRIPE_SECRET_KEY;
-
-export const SubscriptionPage = ({ plans }: ISubscriptionProps): JSX.Element => {
-  const { deviceData } = useDeviceDetect();
-
-  const DashboardLayout = deviceData.isSmallerThanLG
-    ? MobileDashboardLayout
-    : DesktopDashboardLayout;
-
-  console.log(plans);
-
-  return (
-    <DashboardLayout>
-      <Subscription />
-    </DashboardLayout>
-  );
-};
-
-export const getServerSideProps = async (serverSideProps: GetServerSideProps) => {
-  const result = await verifyUserAccess(serverSideProps);
-
-  if (!hasOwnProperty(result, 'profileData') || !hasOwnProperty(result, 'user')) {
-    return result;
-  }
-
-  const stripe = new Stripe(stripeSecretKey, { apiVersion: '2020-08-27' });
-
-  const { data: prices } = await stripe.prices.list();
-
-  const plans = await Promise.all(
-    prices.map(async (_price) => {
-      const product = await stripe.products.retrieve(_price.product.toString());
-
-      return {
-        id: _price.id,
-        name: product.name,
-        currency: _price.currency,
-        price: _price.unit_amount,
-        interval: _price.recurring?.interval,
-      };
-    }),
-  );
-
-  const sortedPlans = plans.sort((a, b) => (a.price || 0) - (b.price || 0));
-
-  return {
-    props: {
-      plans: sortedPlans,
-    },
-  };
-};
