--- src/application/pages/dashboard/subscription/subscription.types.ts
@@ -1,5 +1,0 @@
-import { ISubscriptionPlan } from '@interfaces';
-
-export interface ISubscriptionProps {
-  plans: ISubscriptionPlan[];
-}
