--- src/domain/settings/settings.component.tsx
@@ -1,3 +1,0 @@
-import S from './settings.styles';
-
-export const Settings = (): JSX.Element => <S.StyledWrapper>Settings</S.StyledWrapper>;
