--- src/domain/settings/settings.styles.ts
@@ -1,9 +1,0 @@
-import styled from '@emotion/styled';
-
-const StyledWrapper = styled.div``;
-
-const S = {
-  StyledWrapper,
-};
-
-export default S;
