--- src/domain/settings/settings.types.ts
@@ -1,1 +1,0 @@
-export interface ISettingsProps {}
