--- src/domain/subscription/subscription.component.tsx
@@ -1,3 +1,0 @@
-import S from './subscription.styles';
-
-export const Subscription = (): JSX.Element => <S.StyledWrapper>Subscription</S.StyledWrapper>;
