--- src/domain/subscription/subscription.types.ts
@@ -1,1 +1,0 @@
-export interface ISubscriptionProps {}
