+++ src/infrastructure/i18n/languages/en/dashboard/common.translations.ts
@@ -3,6 +3,4 @@ export default {
   'component.dashboard.sidemenu.option.discover': 'Discover',
   'component.dashboard.sidemenu.option.matches': 'Matches',
   'component.dashboard.sidemenu.option.likes': 'Likes',
-  'component.dashboard.sidemenu.option.subscription': 'Subscription',
-  'component.dashboard.sidemenu.option.settings': 'Settings',
 };
